requirements.txt
fastapi
uvicorn
pymongo
faiss-cpu
redis
celery
deep-translator
python-dotenv
requests
groq
beautifulsoup4
yt_dlp
langchain
transformers
pydantic
pytz
torch
email-validator
pydantic
youtube_transcript_api
youtube-search-python
